Moved SW360ConnectionFactory to adapter package.
Issue: eclipse#426.

This makes it possible to access adapter implementations even if they
are package-local.
