+++ modules/sw360/sw360-workflow/src/test/java/org/eclipse/sw360/antenna/sw360/workflow/SW360ConnectionConfigurationFactoryTest.java
@@ -12,7 +12,7 @@ package org.eclipse.sw360.antenna.sw360.workflow;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.eclipse.sw360.antenna.http.HttpClient;
-import org.eclipse.sw360.antenna.sw360.client.SW360ConnectionFactory;
+import org.eclipse.sw360.antenna.sw360.client.adapter.SW360ConnectionFactory;
 import org.eclipse.sw360.antenna.sw360.client.adapter.SW360Connection;
 import org.eclipse.sw360.antenna.sw360.client.config.SW360ClientConfig;
 import org.junit.Test;
